+++ lib/utilise/augment.rb
@@ -1,10 +1,12 @@
 require 'utilise/augment/bool.rb'
 require 'utilise/augment/matchers.rb'
+require 'utilise/augment/modify.rb'
 require 'utilise/augment/time.rb'
 
 # Extend existing ruby class String
 class String
   include Utilise::Augment::Bool
+  include Utilise::Augment::Modify
 end
 
 # Extend existing ruby class Fixnum