+++ spec/helper.rb
@@ -3,6 +3,3 @@ require 'coveralls'
 Coveralls.wear!
 require 'utilise'
 
-RSpec.configure do |config|
-  config.formatter      = :documentation
-end